Remove unnecessary exclusionlist interface for now
