+++ network-protection/network-protection-impl/src/main/java/com/duckduckgo/networkprotection/impl/config/NetPDefaultConfigProvider.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.networkprotection.impl.config
 
 import com.duckduckgo.di.scopes.VpnScope
-import com.duckduckgo.networkprotection.impl.exclusion.ExclusionList
+import com.duckduckgo.networkprotection.store.NetPExclusionListRepository
 import com.squareup.anvil.annotations.ContributesBinding
 import java.net.Inet4Address
 import java.net.InetAddress
@@ -26,7 +26,7 @@ import javax.inject.Inject
 interface NetPDefaultConfigProvider {
     fun mtu(): Int = 1280
 
-    suspend fun exclusionList(): Set<String> = emptySet()
+    fun exclusionList(): Set<String> = emptySet()
 
     fun fallbackDns(): Set<InetAddress> = emptySet()
 
@@ -46,8 +46,10 @@ interface NetPDefaultConfigProvider {
 data class PcapConfig(val filename: String, val snapLen: Int, val fileSize: Int)
 
 @ContributesBinding(VpnScope::class)
-class RealNetPDefaultConfigProvider @Inject constructor(private val exclusionList: ExclusionList) : NetPDefaultConfigProvider {
-    override suspend fun exclusionList(): Set<String> {
-        return exclusionList.getExclusionAppsList().toSet()
+class RealNetPDefaultConfigProvider @Inject constructor(
+    private val netPExclusionListRepository: NetPExclusionListRepository,
+) : NetPDefaultConfigProvider {
+    override fun exclusionList(): Set<String> {
+        return netPExclusionListRepository.getExcludedAppPackages().toSet()
     }
 }
