+++ network-protection/network-protection-store/src/main/java/com/duckduckgo/networkprotection/store/NetPExclusionListRepository.kt
@@ -27,6 +27,8 @@ interface NetPExclusionListRepository {
 
     fun getManualAppExclusionListFlow(): Flow<List<NetPManuallyExcludedApp>>
 
+    fun getExcludedAppPackages(): List<String>
+
     fun manuallyExcludeApp(packageName: String)
 
     fun manuallyExcludeApps(packageNames: List<String>)
@@ -42,6 +44,13 @@ class RealNetPExclusionListRepository constructor(
     override fun getManualAppExclusionList(): List<NetPManuallyExcludedApp> = exclusionListDao.getManualAppExclusionList()
 
     override fun getManualAppExclusionListFlow(): Flow<List<NetPManuallyExcludedApp>> = exclusionListDao.getManualAppExclusionListFlow()
+
+    override fun getExcludedAppPackages(): List<String> {
+        return getManualAppExclusionList()
+            .filter { !it.isProtected }
+            .sortedBy { it.packageId }
+            .map { it.packageId }
+    }
 
     override fun manuallyExcludeApp(packageName: String) {
         exclusionListDao.insertIntoManualAppExclusionList(NetPManuallyExcludedApp(packageId = packageName, isProtected = false))
